Simplify Scheduler Services
- `MockSpecificationService` was really a `MockSpecificationRepository`, as indicated by its private field that was a repository of Specifications
- The one remaining implementation of `SpecificationService` deferred exclusively to the implementation of its repository, so `SpecificationService` was flattened
- The above applies for `SchedulerService` as well
