--- scheduler-server/src/main/java/gov/nasa/jpl/aerie/scheduler/server/services/LocalSpecificationService.java
@@ -1,23 +1,0 @@
-package gov.nasa.jpl.aerie.scheduler.server.services;
-
-import gov.nasa.jpl.aerie.scheduler.server.exceptions.NoSuchSpecificationException;
-import gov.nasa.jpl.aerie.scheduler.server.exceptions.SpecificationLoadException;
-import gov.nasa.jpl.aerie.scheduler.server.models.Specification;
-import gov.nasa.jpl.aerie.scheduler.server.models.SpecificationId;
-import gov.nasa.jpl.aerie.scheduler.server.remotes.SpecificationRepository;
-
-public record LocalSpecificationService(SpecificationRepository specificationRepository) implements SpecificationService {
-  @Override
-  public Specification getSpecification(final SpecificationId specificationId)
-  throws NoSuchSpecificationException, SpecificationLoadException
-  {
-    return specificationRepository.getSpecification(specificationId);
-  }
-
-  @Override
-  public RevisionData getSpecificationRevisionData(final SpecificationId specificationId)
-  throws NoSuchSpecificationException
-  {
-    return specificationRepository.getSpecificationRevisionData(specificationId);
-  }
-}
