+++ scheduler-server/src/main/java/gov/nasa/jpl/aerie/scheduler/server/services/SpecificationService.java
@@ -4,9 +4,20 @@ import gov.nasa.jpl.aerie.scheduler.server.exceptions.NoSuchSpecificationExcepti
 import gov.nasa.jpl.aerie.scheduler.server.exceptions.SpecificationLoadException;
 import gov.nasa.jpl.aerie.scheduler.server.models.Specification;
 import gov.nasa.jpl.aerie.scheduler.server.models.SpecificationId;
+import gov.nasa.jpl.aerie.scheduler.server.remotes.SpecificationRepository;
+import gov.nasa.jpl.aerie.scheduler.server.remotes.postgres.SpecificationRevisionData;
 
-public interface SpecificationService {
+public record SpecificationService(SpecificationRepository specificationRepository) {
   // Queries
-  Specification getSpecification(SpecificationId specificationId) throws NoSuchSpecificationException, SpecificationLoadException;
-  RevisionData getSpecificationRevisionData(SpecificationId specificationId) throws NoSuchSpecificationException;
+  public Specification getSpecification(final SpecificationId specificationId)
+  throws NoSuchSpecificationException, SpecificationLoadException
+  {
+    return specificationRepository.getSpecification(specificationId);
+  }
+
+  public SpecificationRevisionData getSpecificationRevisionData(final SpecificationId specificationId)
+  throws NoSuchSpecificationException
+  {
+    return specificationRepository.getSpecificationRevisionData(specificationId);
+  }
 }
